migrate new_wallet_from_phrase
