+++ zingolib/src/lightclient/tests.rs
@@ -37,174 +37,6 @@ use crate::wallet::{LightWallet, WalletBase};
 
 use zingoconfig::{ChainType, ZingoConfig};
 
-#[test]
-fn new_wallet_from_phrase() {
-    let temp_dir = tempfile::Builder::new().prefix("test").tempdir().unwrap();
-    let data_dir = temp_dir
-        .into_path()
-        .canonicalize()
-        .unwrap()
-        .to_str()
-        .unwrap()
-        .to_string();
-
-    let wallet_name = format!("{}/{}", &data_dir, "zingo-wallet.dat");
-    let config = ZingoConfig::create_unconnected(ChainType::FakeMainnet, Some(data_dir));
-    let lc = LightClient::new_from_wallet_base(
-        WalletBase::MnemonicPhrase(TEST_SEED.to_string()),
-        &config,
-        0,
-        false,
-    )
-    .unwrap();
-    assert_eq!(
-        format!(
-            "{:?}",
-            LightClient::new_from_wallet_base(
-                WalletBase::MnemonicPhrase(TEST_SEED.to_string()),
-                &config,
-                0,
-                false
-            )
-            .err()
-            .unwrap()
-        ),
-        format!(
-            "{:?}",
-            std::io::Error::new(
-                std::io::ErrorKind::AlreadyExists,
-                format!("Cannot create a new wallet from seed, because a wallet already exists at:\n{:?}", wallet_name),
-            )
-        )
-    );
-
-    // The first t address and z address should be derived
-    Runtime::new().unwrap().block_on(async move {
-        let addresses = lc.do_addresses().await;
-        assert_eq!(
-            "zs1q6xk3q783t5k92kjqt2rkuuww8pdw2euzy5rk6jytw97enx8fhpazdv3th4xe7vsk6e9sfpawfg"
-                .to_string(),
-            addresses[0]["receivers"]["sapling"]
-        );
-        assert_eq!(
-            "t1eQ63fwkQ4n4Eo5uCrPGaAV8FWB2tmx7ui",
-            addresses[0]["receivers"]["transparent"]
-        );
-    });
-}
-
-#[test]
-#[ignore]
-fn new_wallet_from_sapling_esk() {
-    let temp_dir = tempfile::Builder::new().prefix("test").tempdir().unwrap();
-    let data_dir = temp_dir
-        .into_path()
-        .canonicalize()
-        .unwrap()
-        .to_str()
-        .unwrap()
-        .to_string();
-
-    let config = ZingoConfig::create_unconnected(ChainType::FakeMainnet, Some(data_dir));
-    let sk = "secret-extended-key-main1qvpa0qr8qqqqpqxn4l054nzxpxzp3a8r2djc7sekdek5upce8mc2j2z0arzps4zv940qeg706hd0wq6g5snzvhp332y6vhwyukdn8dhekmmsk7fzvzkqm6ypc99uy63tpesqwxhpre78v06cx8k5xpp9mrhtgqs5dvp68cqx2yrvthflmm2ynl8c0506dekul0f6jkcdmh0292lpphrksyc5z3pxwws97zd5els3l2mjt2s7hntap27mlmt6w0drtfmz36vz8pgu7ec0twfrq";
-    // This will panic.
-    // TODO: add Sapling spending key import
-    let lc = LightClient::new_from_wallet_base(WalletBase::Ufvk(sk.to_string()), &config, 0, false)
-        .unwrap();
-    Runtime::new().unwrap().block_on(async move {
-        let addresses = lc.do_addresses().await;
-        assert_eq!(addresses["sapling_addresses"].len(), 1);
-        assert_eq!(addresses["transparent_addresses"].len(), 1);
-        assert_eq!(
-            "zs1q6xk3q783t5k92kjqt2rkuuww8pdw2euzy5rk6jytw97enx8fhpazdv3th4xe7vsk6e9sfpawfg"
-                .to_string(),
-            addresses["sapling_addresses"][0]
-        );
-
-        // New address should be derived from the seed
-        lc.do_new_address("z").await.unwrap();
-
-        let addresses = lc.do_addresses().await;
-        assert_eq!(addresses["sapling_addresses"].len(), 2);
-        assert_ne!(
-            "zs1q6xk3q783t5k92kjqt2rkuuww8pdw2euzy5rk6jytw97enx8fhpazdv3th4xe7vsk6e9sfpawfg"
-                .to_string(),
-            addresses["sapling_addresses"][1]
-        );
-    });
-}
-/*
-#[test]
-fn import_orchard_spending_key() {
-    let temp_dir = tempfile::Builder::new().prefix("test").tempdir().unwrap();
-    let data_dir = temp_dir
-        .into_path()
-        .canonicalize()
-        .unwrap()
-        .to_str()
-        .unwrap()
-        .to_string();
-
-    let config = ZingoConfig::create_unconnected(Network::FakeMainnet, Some(data_dir));
-    Runtime::new().unwrap().block_on(async move {
-        let lc = LightClient::test_new(&config, Some(TEST_SEED.to_string()), 0)
-            .await
-            .unwrap();
-        lc.do_new_address("o").await.unwrap();
-        let new_address = lc
-            .wallet
-            .add_imported_orchard_spending_key(
-                "secret-orchard-sk-main10vj29mt2ezeyc8y5ut6knfcdptg3umdsjk4v8zge6fdmt2kepycqs6j2g8"
-                    .to_string(),
-                0,
-            )
-            .await;
-        assert_eq!(new_address, "Error: Key already exists");
-    });
-}*/
-
-#[test]
-#[ignore]
-fn new_wallet_from_zvk() {
-    let temp_dir = tempfile::Builder::new().prefix("test").tempdir().unwrap();
-    let data_dir = temp_dir
-        .into_path()
-        .canonicalize()
-        .unwrap()
-        .to_str()
-        .unwrap()
-        .to_string();
-
-    let config = ZingoConfig::create_unconnected(ChainType::FakeMainnet, Some(data_dir));
-    let vk = "zxviews1qvpa0qr8qqqqpqxn4l054nzxpxzp3a8r2djc7sekdek5upce8mc2j2z0arzps4zv9kdvg28gjzvxd47ant6jn4svln5psw3htx93cq93ahw4e7lptrtlq7he5r6p6rcm3s0z6l24ype84sgqfrmghu449htrjspfv6qg2zfx2yrvthflmm2ynl8c0506dekul0f6jkcdmh0292lpphrksyc5z3pxwws97zd5els3l2mjt2s7hntap27mlmt6w0drtfmz36vz8pgu7ecrxzsls";
-    // This will panic
-    // TODO: add Sapling FVKs import
-    let lc = LightClient::new_from_wallet_base(WalletBase::Ufvk(vk.to_string()), &config, 0, false)
-        .unwrap();
-
-    Runtime::new().unwrap().block_on(async move {
-        let addresses = lc.do_addresses().await;
-        assert_eq!(addresses["sapling_addresses"].len(), 1);
-        assert_eq!(addresses["transparent_addresses"].len(), 1);
-        assert_eq!(
-            "zs1q6xk3q783t5k92kjqt2rkuuww8pdw2euzy5rk6jytw97enx8fhpazdv3th4xe7vsk6e9sfpawfg"
-                .to_string(),
-            addresses["sapling_addresses"][0]
-        );
-
-        // New address should be derived from the seed
-        lc.do_new_address("z").await.unwrap();
-
-        let addresses = lc.do_addresses().await;
-        assert_eq!(addresses["sapling_addresses"].len(), 2);
-        assert_ne!(
-            "zs1q6xk3q783t5k92kjqt2rkuuww8pdw2euzy5rk6jytw97enx8fhpazdv3th4xe7vsk6e9sfpawfg"
-                .to_string(),
-            addresses["sapling_addresses"][1]
-        );
-    });
-}
-
 apply_scenario! {sapling_incoming_sapling_outgoing 1}
 async fn sapling_incoming_sapling_outgoing(scenario: NBlockFCBLScenario) {
     let NBlockFCBLScenario {
